12/10 test 환경 세팅, CI CD 테스트 1차

+++ models/index.js
@@ -9,33 +9,36 @@ const env = process.env.NODE_ENV || "development";
 const config = require(__dirname + "/../config/config.js")[env];
 const db = {};
 
-let sequelize;
-if (config.use_env_variable) {
-  sequelize = new Sequelize(process.env[config.use_env_variable], config);
-} else {
-  sequelize = new Sequelize(config.database, config.username, config.password, config);
-}
+if(process.env.NODE_ENV !== 'test'){ // test 상황 조건문 여기서부터-
+    let sequelize;
+    if (config.use_env_variable) {
+      sequelize = new Sequelize(process.env[config.use_env_variable], config);
+    } else {
+      sequelize = new Sequelize(config.database, config.username, config.password, config);
+    }
 
-fs.readdirSync(__dirname)
-  .filter((file) => {
-    return file.indexOf(".") !== 0 && file !== basename && file.slice(-3) === ".js";
-  })
-  .forEach((file) => {
-    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
-    db[model.name] = model;
-  });
 
-Object.keys(db).forEach((modelName) => {
-  if (db[modelName].associate) {
-    db[modelName].associate(db);
-  }
-});
+    fs.readdirSync(__dirname)
+      .filter((file) => {
+        return file.indexOf(".") !== 0 && file !== basename && file.slice(-3) === ".js";
+      })
+      .forEach((file) => {
+        const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+        db[model.name] = model;
+      });
 
-db.sequelize = sequelize;
-db.Sequelize = Sequelize;
+    Object.keys(db).forEach((modelName) => {
+      if (db[modelName].associate) {
+        db[modelName].associate(db);
+      }
+    });
 
-sequelize.sync({ force: false });
+    db.sequelize = sequelize;
+    db.Sequelize = Sequelize;
 
-// force true 시 디비 날려버리고 새로만듭니다! 첫실행 후 꼭 false로 바꿔주기!
+    // force true 시 디비 날려버리고 새로만듭니다! 첫실행 후 꼭 false로 바꿔주기!
+    sequelize.sync({ force: false });
+
+} //test 상황 조건문 여기까지 감싸주기
 
 module.exports = db;
